Updated trading.py
Essentially all of it is done except for the main function that calls to get the arrays representing the RNN and stock data. I am going to wait for jeremy to finish before implementing this part.

In addition, the I just want to make sure that we are getting the stock data from the right place.

We might want to write another script where we compare sentiment analysis to the stock data (but that may be a part of the regression stuff anyways).