Updated wordcloud and added new histogram visualization
New Histogram visualization works on the entire tweet database rather than using the compiled preprocessed tweet sentiment that we were calculating before.

This may be relevant since it is the raw data and we can compare it to the sentiment data to show that we have still preserved the relevant part of the data while we were passing it in. The only difference is that there is now a greater deviation due to popularity of tweets being included in sentiment score.